Enemy and player systems updated.
I have added a simple system that makes enemies detect players within a certain distance and start to follow them with different speeds/ranges based on enemy names. Also added an autoload to create the player in any scene that is not excluded. Needs to update stats from a file for both player and enemies so I can also use them in fight scene.