Changed how combat scene appears.
No longer switching scenes. Just going to hide and unhide the scene by attaching it to the player. Stops all functions of everything else when combat starts, just needs to unfreeze it after. Have basic UI layout of combat, but need to do actual logic for it and make enemies appear.