SceneManager Global Working
Fixed and confirmed that the scene_manager global/auto_load function properly changes the scene to the next scene as well as confirming that calling get_tree() afterwards does in fact return the new tree as intended. A debug function called _on_scene_changed() and get_tree().connect() snippet is in the scene_manager which can be uncommented to display the current working scene after a scene change to ensure that the player is in the proper scene if necessary.