Comments and Tilemap Fixes
Added comments explaining the purpose of base_scene.gd, SceneManager.gd, and scene_trigger.gd. Fixed some improper collision interactions with the pillar tops in Room11 as well as the problem of the player being able to see the edge of the tilemap when positioned in certain spots of Rooms03, 04, 06, and 07.